adjust flags to be json
